refactor(app): stripe entities
