+++ openmeter/appstripe/entity/stripe.go
@@ -12,53 +12,11 @@ import (
 	appstripecustomerdb "github.com/openmeterio/openmeter/openmeter/ent/db/appstripecustomer"
 )
 
-var (
-	StripeMarketplaceListing = appentitybase.MarketplaceListing{
-		Type:        appentitybase.AppTypeStripe,
-		Name:        "Stripe",
-		Description: "Stripe is a payment processing platform.",
-		IconURL:     "https://stripe.com/favicon.ico",
-		Capabilities: []appentitybase.Capability{
-			StripeCollectPaymentCapability,
-			StripeCalculateTaxCapability,
-			StripeInvoiceCustomerCapability,
-		},
-	}
-
-	StripeCollectPaymentCapability = appentitybase.Capability{
-		Type:        appentitybase.CapabilityTypeCollectPayments,
-		Key:         "stripe_collect_payment",
-		Name:        "Payment",
-		Description: "Process payments",
-	}
-
-	StripeCalculateTaxCapability = appentitybase.Capability{
-		Type:        appentitybase.CapabilityTypeCalculateTax,
-		Key:         "stripe_calculate_tax",
-		Name:        "Calculate Tax",
-		Description: "Calculate tax for a payment",
-	}
-
-	StripeInvoiceCustomerCapability = appentitybase.Capability{
-		Type:        appentitybase.CapabilityTypeInvoiceCustomers,
-		Key:         "stripe_invoice_customer",
-		Name:        "Invoice Customer",
-		Description: "Invoice a customer",
-	}
-)
-
-// if err != nil {
-// 	return nil, fmt.Errorf("failed to register marketplace listing: %w", err)
-// }
-
 // App represents an installed Stripe app
 type App struct {
 	appentitybase.AppBase
 
-	// TODO: cycle dependencies
 	Client *entdb.Client
-	// AppService app.Service
-	// AppStripeService appstripe.Service
 	// BillingService   billing.Service
 
 	StripeAccountId string `json:"stripeAccountId"`
@@ -135,6 +93,7 @@ func (a App) ValidateCustomer(ctx context.Context, customer *customerentity.Cust
 	return nil
 }
 
+// CustomerAppData represents the Stripe associated data for an app used by a customer
 type CustomerAppData struct {
 	StripeCustomerID string `json:"stripeCustomerId"`
 }
@@ -142,91 +101,6 @@ type CustomerAppData struct {
 func (d CustomerAppData) Validate() error {
 	if d.StripeCustomerID == "" {
 		return errors.New("stripe customer id is required")
-	}
-
-	return nil
-}
-
-type CreateAppStripeInput struct {
-	Namespace       string
-	Name            string
-	Description     string
-	StripeAccountID string
-	Livemode        bool
-}
-
-func (i CreateAppStripeInput) Validate() error {
-	if i.Namespace == "" {
-		return errors.New("namespace is required")
-	}
-
-	if i.Name == "" {
-		return errors.New("name is required")
-	}
-
-	if i.Description == "" {
-		return errors.New("description is required")
-	}
-
-	if i.StripeAccountID == "" {
-		return errors.New("stripe account id is required")
-	}
-
-	return nil
-}
-
-type UpsertStripeCustomerDataInput struct {
-	AppID            appentitybase.AppID
-	CustomerID       customerentity.CustomerID
-	StripeCustomerID string
-}
-
-func (i UpsertStripeCustomerDataInput) Validate() error {
-	if err := i.AppID.Validate(); err != nil {
-		return fmt.Errorf("error validating app id: %w", err)
-	}
-
-	if err := i.CustomerID.Validate(); err != nil {
-		return fmt.Errorf("error validating customer id: %w", err)
-	}
-
-	if i.AppID.Namespace != i.CustomerID.Namespace {
-		return errors.New("app and customer must be in the same namespace")
-	}
-
-	if i.StripeCustomerID == "" {
-		return errors.New("stripe customer id is required")
-	}
-
-	return nil
-}
-
-type DeleteStripeCustomerDataInput struct {
-	AppID      *appentitybase.AppID
-	CustomerID customerentity.CustomerID
-}
-
-func (i DeleteStripeCustomerDataInput) Validate() error {
-	if i.CustomerID.ID == "" {
-		return errors.New("customer id is required")
-	}
-
-	if i.CustomerID.Namespace == "" {
-		return errors.New("customer namespace is required")
-	}
-
-	if i.AppID != nil {
-		if i.AppID.ID == "" {
-			return errors.New("app id is required")
-		}
-
-		if i.AppID.Namespace == "" {
-			return errors.New("app namespace is required")
-		}
-
-		if i.AppID.Namespace != i.CustomerID.Namespace {
-			return errors.New("app and customer must be in the same namespace")
-		}
 	}
 
 	return nil
